+++ src/builtin-plugins/todo/components/TodoMoreSettings.tsx
@@ -92,7 +92,7 @@ export function MoreSettingsDialog({
             <DialogTitle>更多配置</DialogTitle>
           </DialogHeader>
 
-          <DialogContent className="no-scrollbar flex max-h-[min(380px,60vh)] flex-col gap-5">
+          <DialogContent className="flex flex-col gap-5">
             {showTags && (
               <TagDraftList
                 items={tags}