+++ src/components/ImagePreviewDialog.tsx
@@ -39,7 +39,7 @@ export function ImagePreviewDialog({
         <DialogHeader>
           <DialogTitle className="truncate">图片预览</DialogTitle>
         </DialogHeader>
-        <DialogContent className="flex min-h-0 flex-1 flex-col overflow-hidden p-3 pt-0">
+        <DialogContent className="flex min-h-0 flex-1 flex-col overflow-hidden p-3 pt-0" scrollable={false}>
           {image && <ImagePreviewViewport src={image.src} alt={image.alt} />}
         </DialogContent>
       </DialogPanel>